Modificação no comando de atualização, adição de .gitignore e umas especificações do README.md

+++ init.lua
@@ -22,14 +22,27 @@ require("configs.config")
 require("configs.keymaps")
 require("plugins.plugins_config.core.lsp")
 
--- Comando para atualização
 vim.api.nvim_create_user_command('VVUpdate', function()
     os.execute("cd ~/.config/nvim")
+
+    local handle = io.popen("git fetch --dry-run 2>&1")
+    local fetch_result = handle:read("*a")
+    handle:close()
+
+    if fetch_result == "" then
+        print("O VoidVim já está atualizada.")
+        os.execute("cd ~")
+        return
+    end
+
     os.execute("git pull origin")
+
     vim.cmd('Lazy sync')
     os.execute("rm -rf LICENCE README.md")
+    print("O VoidVim foi atualizado com sucesso!")
+    print("Bom código")
+
     os.execute("cd ~")
-end,
-{})
+end, {})
 
 -- Obs: Os comentários só são uma "ajuda" pra quem quer analisar o código, ou os que querem fazer uma fork do projeto, caso queira que fazer suas próprias configurações dê uma olhada na pasta ~/.config/nvim/lua/opt faça as suas configurações do seu jeito
